renamed JoinManager in QueryManager, code refactoring
